refactor: move testtoggle
